feat: control de permisos por rol, TTS es-AR, auth y mejoras
- Control de permisos: config centralizada, ProtectedRoute, filtrado sidebar/nav
- Roles: admin, supervisor, operador, visualizador con permisos granulares
- Auth: login, sesiones, JWT, authService, rutas protegidas
- TTS: voces es-AR prioritarias, es-ES como Ãºltimo recurso, index lang es-AR
- Accesibilidad y mejores prÃ¡cticas en AccessDenied y ProtectedRoute

+++ apps/backend/src/services/wazePollingService.ts
@@ -189,7 +189,7 @@ export class WazePollingService {
                 CREATE TABLE IF NOT EXISTS waze_tvt_metrics (
                     id SERIAL PRIMARY KEY,
                     polygon_id VARCHAR(50) NOT NULL,
-                    wazers_count INTEGER DEFAULT 0,
+                    wazers_count NUMERIC(10, 2) DEFAULT 0,
                     jam_level_counts JSONB,
                     length_of_jams JSONB,
                     update_time TIMESTAMPTZ,
